server/requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
docopt==0.6.2
enum34==1.1.6
httpserver==1.1.0
idna==2.7
ipaddress==1.0.22
pycparser==2.18
pyOpenSSL==18.0.0
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
